refactor: 방 조회 담당 도메인 변경 및 return type 통일
