+++ src/main/java/com/snowhite/server/service/RoomService.java
@@ -3,11 +3,19 @@ package com.snowhite.server.service;
 import com.snowhite.server.domain.session.Game;
 import com.snowhite.server.domain.session.Player;
 import com.snowhite.server.domain.session.Room;
+import com.snowhite.server.payload.ApiResponse;
+import com.snowhite.server.web.dto.web.response.GetRoomResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.ReactiveRedisTemplate;
+import org.springframework.data.redis.core.ScanOptions;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
+import org.springframework.web.reactive.function.server.ServerRequest;
+import org.springframework.web.reactive.function.server.ServerResponse;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -20,11 +28,11 @@ public class RoomService {
     private final ReactiveRedisTemplate<String, Game> reactiveRedisTemplateForGame;
     private final ReactiveRedisTemplate<String, Room> reactiveRedisTemplateForRoom;
 
+
     public Mono<Long> startGameByRoomId(Long roomId) {
 
         Mono<Room> roomForStart = reactiveRedisTemplateForRoom.opsForValue().get(ROOM_PREFIX + roomId);
 
-        // Game 객체 생성 후 Redis 저장
         Mono<Long> gameId = roomForStart
                 .flatMap(room -> {
                     List<Player> players = room.getUsers().stream()
@@ -38,6 +46,26 @@ public class RoomService {
                 });
 
         return gameId;
+    }
+
+    public Mono<GetRoomResponse> getRooms() {
+
+        return scanRoomKeys()
+                .flatMap(roomId -> reactiveRedisTemplateForRoom.opsForValue().get(roomId))
+                .collectList()
+                .flatMap(rooms -> {
+                    if (rooms == null) {
+                        rooms = Collections.emptyList();
+                    }
+                    GetRoomResponse.of(rooms);
+                    return Mono.just(GetRoomResponse.of(rooms));
+                });
+    }
+
+    public Flux<String> scanRoomKeys() {
+        ScanOptions options =
+                ScanOptions.scanOptions().match("room:*").count(1000).build();
+        return reactiveRedisTemplateForRoom.scan(options);
     }
 
 }
