+++ src/main/java/com/snowhite/server/web/controller/RoomController.java
@@ -2,11 +2,9 @@ package com.snowhite.server.web.controller;
 
 import com.snowhite.server.payload.ApiResponse;
 import com.snowhite.server.service.RoomService;
+import com.snowhite.server.web.dto.web.response.GetRoomResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
 
 @RestController
@@ -23,5 +21,15 @@ public class RoomController {
 
         Mono<ApiResponse<Long>> result = gameId.map(ApiResponse::onSuccess);
         return result;
+    }
+
+    @GetMapping()
+    public Mono<ApiResponse<GetRoomResponse>> getRooms() {
+
+        Mono<GetRoomResponse> getRoomResponse = roomService.getRooms();
+
+        Mono<ApiResponse<GetRoomResponse>> result = getRoomResponse.map(ApiResponse::onSuccess);
+        return result;
+
     }
 }
