+++ src/main/java/com/snowhite/server/web/dto/web/response/GetRoomResponse.java
@@ -1,0 +1,13 @@
+package com.snowhite.server.web.dto.web.response;
+
+import com.snowhite.server.domain.session.Room;
+
+import java.util.List;
+
+public record GetRoomResponse(
+        List<Room> roomList
+) {
+    public static GetRoomResponse of(List<Room> roomList) {
+        return new GetRoomResponse(roomList);
+    }
+}
